refactored code to match cobra scafolding
