+++ cmd/ec2_test.go
